Prepare what() and where() messages in Exception constructor
Because of giving more flexibility on generating what() and where()
messages, add Exception::{what,where}Str, which are prepared in
constructor in Exception::prepare{What,Where}() functions (which
are protected), respectively.

The other way to prepare messages of what() and where() is in the
functions themselves; however, this idea make things a bit
difficult and complicated, because their return value is const char
*, and not std::string. In this case, creating a std::string in
these functions and return their c_str() doesn't help, as they are
local to the function.