Implement Exception::has*() functions
Also, mark all getters and has*()s final (i.e. could not be
overriden). The reason is to avoid complexity. Returning something
else from the getters does not mean anything, the return values
are clear, so they should not be changed. Furthermore, if there
will be a need not being final for some reason, the final
qualifiers will be removed. Note, explicit is better than implicit.